+++ lib/core/base_on.dart
@@ -1,1 +1,5 @@
+// Copyright 2021 All wao_ui authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 class BaseOn {}