+++ lib/core/byte_file.dart
@@ -1,6 +1,8 @@
-import 'dart:typed_data';
+// Copyright 2021 All wao_ui authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
-import 'package:dio/dio.dart';
+import 'dart:typed_data';
 
 class ByteFile {
   late Uint8List bytes;