+++ lib/core/theme/element/swatch_css.dart
@@ -1,4 +1,6 @@
-import 'package:flutter/material.dart';
+// Copyright 2021 All wao_ui authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 import '../../../wao_ui.dart';
 import '../../base_style.dart';