+++ lib/core/utils/collect_util.dart
@@ -1,4 +1,6 @@
-import 'dart:collection';
+// Copyright 2021 All wao_ui authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 import 'package:flutter/material.dart';