+++ lib/src/basic/w_button.dart
@@ -1,4 +1,6 @@
-import 'dart:math';
+// Copyright 2021 All wao_ui authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 import 'package:flutter/material.dart';
 import 'package:wao_ui/core/base_on.dart';
@@ -8,7 +10,6 @@ import 'package:wao_ui/core/base_mixins.dart';
 import 'package:wao_ui/src/basic/cfg_global.dart';
 import 'package:wao_ui/src/others/w_spin.dart';
 import 'package:wao_ui/wao_ui.dart';
-import 'package:bitsdojo_window/bitsdojo_window.dart';
 
 class WButton
     extends WStatefulWidget<WButtonOn, WButtonProp, WButtonSlot, WButtonStyle> {