+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+<!-- 
+  Copyright 2021 All wao_ui authors. All rights reserved.
+  Use of this source code is governed by a BSD-style license that can be
+  found in the LICENSE file.
+ -->
 
-* TODO: Describe initial release.
+## 0.0.1-dev.1
+
+* Test publish